requirements.txt
Flask==3.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
numpy==1.26.0
pandas==1.4.2